moved transform function  into etl dir and updated existing tests
